Nav cleanup: remove Changes, merge Manager onto Reviews/Allowlist row, rename "Allowlist"
Three related top-level navigation changes (#385):

1. Remove the unused Changes feature end-to-end — the sidebar button,
   SummaryBoardView, the `crow summary` CLI command + its `get-summary`
   RPC handler, GitManager.summarizeCommits and its summary-only helpers
   (parseCommitLog, remoteInfo/RemoteInfo, host/slug/commitURLPrefix),
   the CommitInfo/RepoCommitSummary models, the summaryRepos config and
   its "Changes Summary" Settings section, related AppState/AppDelegate
   wiring, and the summary-only tests. No other feature depended on this
   chain (discoverRepos/resolveDefaultBranch are kept).

2. Consolidate the sidebar so Manager, Reviews, and Allowlist share one
   row (ManagerReviewsAllowListRow); delete the standalone
   ManagerSidebarRow.

3. Rename the "Allow List" label to the single word "Allowlist".


Crow-Session: 5260534B-3E30-4D76-A4C0-7F5CB9DC5249
