+++ Packages/CrowGit/Sources/CrowGit/GitManager.swift
@@ -235,201 +235,6 @@ public actor GitManager {
         }
     }
 
-    /// Collect commits across discovered repos within a time window and group
-    /// them by repo. Deterministic digest — no LLM.
-    ///
-    /// - `since`/`until` are passed straight to git, which parses flexible date
-    ///   strings ("1 week ago", "2026-05-01"); `until` defaults to now when nil.
-    /// - `includeRepos`, when non-nil, scopes the scan to repos whose `org/repo`
-    ///   slug (derived from the `origin` remote, matched case-insensitively) is
-    ///   in the set (an empty set yields nothing); `nil` includes every
-    ///   discovered repo. Bare names are ambiguous across orgs, so the scope key
-    ///   is the full slug; a repo with no parseable remote can't match a scoped
-    ///   set and is dropped.
-    /// - Before scanning each repo we `git fetch --no-tags origin <defaultBranch>`
-    ///   and base the digest on `origin/<defaultBranch>` (the latest remote
-    ///   default branch) so it reflects what landed on the remote regardless of
-    ///   local state. The default branch is resolved from `origin/HEAD`; if that
-    ///   can't be resolved we fall back to `--all` (commits on every branch).
-    ///   `--no-merges` keeps diffstats from double-counting merge commits.
-    /// - Repos sharing an `org/repo` slug are deduped (first seen wins) so each
-    ///   configured repo appears at most once.
-    /// - Repos with zero commits in the window are omitted; a repo whose
-    ///   `git log` errors (empty/detached) is skipped rather than failing the
-    ///   whole scan.
-    ///
-    /// Runs sequentially: `run` is actor-isolated and blocks on
-    /// `waitUntilExit`, so a task group would serialize on the actor anyway.
-    public func summarizeCommits(since: String, until: String?, includeRepos: Set<String>? = nil) async throws -> [RepoCommitSummary] {
-        let repos = try await discoverRepos()
-        // Configured scope keys are `org/repo` slugs; compare case-insensitively
-        // since git hosts treat org/repo case-insensitively. nil = include all.
-        let wantedSlugs = includeRepos.map { Set($0.map { $0.lowercased() }) }
-        var summaries: [RepoCommitSummary] = []
-        var seenSlugs = Set<String>()
-        for repo in repos {
-            // Derive org/repo + the hosted-commit-page prefix from the remote.
-            // When scoping, skip repos that don't match (or have no remote).
-            let remote = await remoteInfo(forRepoAt: repo.path)
-            if let wantedSlugs {
-                guard let slug = remote?.slug, wantedSlugs.contains(slug.lowercased()) else { continue }
-            }
-            // Dedupe by slug so a repo checked out in two workspaces appears once.
-            if let slug = remote?.slug.lowercased() {
-                guard seenSlugs.insert(slug).inserted else { continue }
-            }
-            // Base the digest on the latest remote default branch: fetch it, then
-            // log `origin/<default>`. Fall back to `--all` if it can't be resolved.
-            let defaultBranch = try? await resolveDefaultBranch(repoPath: repo.path)
-            if let defaultBranch {
-                _ = try? await run(["git", "-C", repo.path, "fetch", "--no-tags", "origin", defaultBranch])
-            }
-            let logRange = defaultBranch.map { "origin/\($0)" } ?? "--all"
-            var args = [
-                "git", "-C", repo.path, "log", logRange, "--no-merges",
-                "--since=\(since)",
-            ]
-            if let until { args.append("--until=\(until)") }
-            args.append(contentsOf: [
-                "--date=iso-strict",
-                "--pretty=format:%x1e%H%x1f%h%x1f%an%x1f%ae%x1f%aI%x1f%s",
-                "--numstat",
-            ])
-
-            let output: String
-            do {
-                output = try await run(args)
-            } catch {
-                // Empty repo, detached HEAD, etc. — skip, don't abort the scan.
-                continue
-            }
-
-            let commits = Self.parseCommitLog(output)
-            if !commits.isEmpty {
-                summaries.append(RepoCommitSummary(
-                    repo: repo.name, path: repo.path, workspace: repo.workspace,
-                    commits: commits, commitURLPrefix: remote?.commitURLPrefix
-                ))
-            }
-        }
-        return summaries
-    }
-
-    /// Parsed `origin` remote for a repo: the `org/repo` slug and the
-    /// hosted-commit-page URL prefix. Used both to scope the summary scan to an
-    /// `org/repo` allowlist and to make commits clickable in the UI.
-    struct RemoteInfo: Sendable {
-        let slug: String        // "org/repo" (or nested "group/sub/repo")
-        let commitURLPrefix: String   // "https://host/slug/commit/" or ".../-/commit/"
-    }
-
-    /// Resolve the `origin` remote of the repo at `path` into a slug + commit
-    /// URL prefix. Handles SSH (`git@host:org/repo`) and HTTPS
-    /// (`https://host/org/repo`), strips a trailing `.git`. GitLab hosts use the
-    /// `/-/commit/` path; everything else (GitHub) uses `/commit/`. Returns nil
-    /// when the repo has no remote or it can't be parsed.
-    func remoteInfo(forRepoAt path: String) async -> RemoteInfo? {
-        guard let raw = try? await run(["git", "-C", path, "remote", "get-url", "origin"]) else { return nil }
-        let url = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !url.isEmpty else { return nil }
-        let host = Self.host(fromRemote: url)
-        let slug = Self.slug(fromRemote: url)
-        guard !host.isEmpty, !slug.isEmpty else { return nil }
-        return RemoteInfo(slug: slug, commitURLPrefix: Self.commitURLPrefix(host: host, slug: slug))
-    }
-
-    /// Build the hosted-commit-page prefix for a host + slug. GitLab puts commits
-    /// under `/-/commit/`; GitHub (and the default) uses `/commit/`.
-    ///
-    /// Provider detection is a substring check on the host. A self-hosted GitLab
-    /// whose hostname doesn't contain "gitlab" (e.g. `git.company.com`) falls
-    /// through to the GitHub-style `/commit/` path and would produce a wrong URL;
-    /// that's an accepted limitation short of probing the remote.
-    static func commitURLPrefix(host: String, slug: String) -> String {
-        let commitPath = host.lowercased().contains("gitlab") ? "/-/commit/" : "/commit/"
-        return "https://\(host)/\(slug)\(commitPath)"
-    }
-
-    /// Extract the host ("github.com", "gitlab.example.com") from a git remote
-    /// URL. Handles SSH (`git@host:org/repo`) and HTTPS (`https://host/...`).
-    static func host(fromRemote url: String) -> String {
-        // SSH: git@host:org/repo — host runs up to the first colon and excludes
-        // '/' so a path-ish remote (git@host/extra:org/repo) can't fold the path
-        // into the host (matches the HTTPS branch's intent).
-        if let range = url.range(of: #"^[^@]+@([^:/]+):"#, options: .regularExpression) {
-            let match = String(url[range])
-            if let at = match.firstIndex(of: "@"), let colon = match.lastIndex(of: ":") {
-                return String(match[match.index(after: at)..<colon])
-            }
-        }
-        // HTTPS: https://host/...
-        if let range = url.range(of: #"^https?://([^/]+)/"#, options: .regularExpression) {
-            let match = String(url[range])
-            return match
-                .replacingOccurrences(of: #"^https?://"#, with: "", options: .regularExpression)
-                .trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        }
-        return ""
-    }
-
-    /// Extract the project slug ("org/repo", "group/sub/repo") from a git remote
-    /// URL. Handles SSH and HTTPS, preserves nested-group paths, strips trailing
-    /// `.git`. Returns "" when the URL can't be parsed.
-    static func slug(fromRemote url: String) -> String {
-        var trimmed = url.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmed.hasSuffix(".git") { trimmed = String(trimmed.dropLast(4)) }
-        // SSH: git@host:org/repo or user@host:group/sub/repo
-        if let range = trimmed.range(of: #"^[^@/\s]+@[^:/\s]+:"#, options: .regularExpression) {
-            return String(trimmed[range.upperBound...]).trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        }
-        // HTTPS: https://host/org/repo
-        if let range = trimmed.range(of: #"^https?://[^/]+/"#, options: .regularExpression) {
-            return String(trimmed[range.upperBound...]).trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-        }
-        return ""
-    }
-
-    /// Parse the `--pretty`/`--numstat` output of the `summarizeCommits` git
-    /// command. Records are delimited by RS (`\u{1e}`), header fields by US
-    /// (`\u{1f}`); numstat lines (`<ins>\t<del>\t<path>`) follow each header.
-    static func parseCommitLog(_ output: String) -> [CommitInfo] {
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = [.withInternetDateTime]
-        var commits: [CommitInfo] = []
-
-        for record in output.components(separatedBy: "\u{1e}") {
-            let trimmed = record.trimmingCharacters(in: .newlines)
-            guard !trimmed.isEmpty else { continue }
-
-            let lines = trimmed.components(separatedBy: "\n")
-            let fields = lines[0].components(separatedBy: "\u{1f}")
-            guard fields.count >= 6 else { continue }
-
-            var files = 0, insertions = 0, deletions = 0
-            for line in lines.dropFirst() where !line.isEmpty {
-                let cols = line.components(separatedBy: "\t")
-                guard cols.count >= 3 else { continue }
-                files += 1
-                // Binary files report "-" for ins/del; treat as 0.
-                insertions += Int(cols[0]) ?? 0
-                deletions += Int(cols[1]) ?? 0
-            }
-
-            commits.append(CommitInfo(
-                hash: fields[0],
-                shortHash: fields[1],
-                authorName: fields[2],
-                authorEmail: fields[3],
-                date: formatter.date(from: fields[4]) ?? Date(timeIntervalSince1970: 0),
-                subject: fields[5],
-                filesChanged: files,
-                insertions: insertions,
-                deletions: deletions
-            ))
-        }
-        return commits
-    }
-
     /// Run a git command, returning stdout on success.
     private func run(_ args: [String]) async throws -> String {
         let process = Process()
