+++ Packages/CrowUI/Sources/CrowUI/MainContentView.swift
@@ -20,8 +20,6 @@ public struct MainContentView: View {
                 AllowListView(appState: appState)
             } else if appState.selectedSessionID == AppState.reviewBoardSessionID {
                 ReviewBoardView(appState: appState)
-            } else if appState.selectedSessionID == AppState.summaryBoardSessionID {
-                SummaryBoardView(appState: appState)
             } else if let session = appState.selectedSession {
                 SessionDetailView(session: session, appState: appState)
             } else {
